examples/requirements.txt
uvicorn
python-gql
starlette-graphql
databases
sqlalchemy